pyproject.toml
[project]
name = "my-robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.6.4",
    "openai>=2.14.0",
]
